Move form fill/cwc logic into concerns
