+++ app/concerns/form_filling.rb
@@ -1,0 +1,258 @@
+module FormFilling
+  extend ActiveSupport::Concern
+
+  included do
+    attr_accessor :persist_session
+    alias_method :persist_session?, :persist_session
+  end
+
+
+  def fill_out_form f={}, ct = nil, session: nil, action: nil, &block
+    status_fields = {
+      congress_member: self,
+      status: "success",
+      extra: {}
+    }
+    status_fields[:campaign_tag] = ct unless ct.nil?
+
+    success_hash = fill_out_form_with_capybara f, session, starting_action: action, &block
+
+    unless success_hash[:success]
+      status_fields[:status] = success_hash[:exception] ? "error" : "failure"
+      status_fields[:extra][:screenshot] = success_hash[:screenshot]
+
+      if success_hash[:exception]
+        message = YAML.load(success_hash[:exception].message)
+
+        if message.is_a?(Hash) and message.include?(:screenshot)
+          status_fields[:extra][:screenshot] = message[:screenshot]
+        end
+      end
+    end
+
+    fill_status = FillStatus.create(status_fields)
+    fill_status.save if RECORD_FILL_STATUSES
+
+    fill_status
+  end
+
+  def fill_out_form!(f={}, ct=nil, &block)
+    fill_out_form(f, ct, &block)[0] or raise FillError.new
+  end
+
+  DEFAULT_FIND_WAIT_TIME = 5  
+
+  def fill_out_form_with_capybara f, session=nil, starting_action: nil
+    session ||= Capybara::Session.new(:poltergeist)
+    session.driver.options[:js_errors] = false
+    session.driver.options[:phantomjs_options] = ['--ssl-protocol=TLSv1']
+    form_fill_log(f, "begin")
+
+    begin
+      actions = self.actions.order(:step)
+
+      if starting_action
+        actions = actions.drop_while{ |a| a.id != starting_action.id }
+      end
+
+      actions.each do |a|
+        form_fill_log(f, %(#{a.action}(#{a.selector.inspect+", " if a.selector.present?}#{a.value.inspect})))
+
+        case a.action
+        when "visit"
+          session.visit(a.value)
+        when "wait"
+          sleep a.value.to_i
+        when "fill_in"
+          if a.value.starts_with?("$")
+            if a.value == "$CAPTCHA_SOLUTION"
+              location = CAPTCHA_LOCATIONS.keys.include?(bioguide_id) ? CAPTCHA_LOCATIONS[bioguide_id] : session.driver.evaluate_script('document.querySelector("' + a.captcha_selector.gsub('"', '\"') + '").getBoundingClientRect();')
+              url = self.class::save_captcha_and_store_poltergeist session, location["left"], location["top"], location["width"], location["height"]
+
+              yield(url, session, a) unless f["$CAPTCHA_SOLUTION"]
+              session.find(a.selector).set(f["$CAPTCHA_SOLUTION"])
+            else
+              if a.options
+                options = YAML.load a.options
+                if options.include? "max_length"
+                  f[a.value] = f[a.value][0...(0.95 * options["max_length"]).floor] unless f[a.value].nil?
+                end
+              end
+              session.find(a.selector).set(f[a.value].gsub("\t","    ")) unless f[a.value].nil?
+            end
+          else
+            session.find(a.selector).set(a.value) unless a.value.nil?
+          end
+        when "select"
+          begin
+            session.within a.selector do
+              if f[a.value].nil?
+                unless PLACEHOLDER_VALUES.include? a.value
+                  begin
+                    elem = session.find('option[value="' + a.value.gsub('"', '\"') + '"]')
+                  rescue Capybara::Ambiguous
+                    elem = session.first('option[value="' + a.value.gsub('"', '\"') + '"]')
+                  rescue Capybara::ElementNotFound
+                    begin
+                      elem = session.find('option', text: Regexp.compile("^" + Regexp.escape(a.value) + "(\\W|$)"))
+                    rescue Capybara::Ambiguous
+                      elem = session.first('option', text: Regexp.compile("^" + Regexp.escape(a.value) + "(\\W|$)"))
+                    end
+                  end
+                  elem.select_option
+                end
+              else
+                begin
+                  elem = session.find('option[value="' + f[a.value].gsub('"', '\"') + '"]')
+                rescue Capybara::Ambiguous
+                  elem = session.first('option[value="' + f[a.value].gsub('"', '\"') + '"]')
+                rescue Capybara::ElementNotFound
+                  begin
+                    elem = session.find('option', text: Regexp.compile("^" + Regexp.escape(f[a.value]) + "(\\W|$)"))
+                  rescue Capybara::Ambiguous
+                    elem = session.first('option', text: Regexp.compile("^" + Regexp.escape(f[a.value]) + "(\\W|$)"))
+                  end
+                end
+                elem.select_option
+              end
+            end
+          rescue Capybara::ElementNotFound => e
+            raise e, e.message unless a.options == "DEPENDENT"
+          end
+        when "click_on"
+          session.find(a.selector).click
+        when "find"
+          wait_val = DEFAULT_FIND_WAIT_TIME
+          if a.options
+            options_hash = YAML.load a.options
+            wait_val = options_hash['wait'] || DEFAULT_FIND_WAIT_TIME
+          end
+          if a.value.nil?
+            session.find(a.selector, wait: wait_val)
+          else
+            session.find(a.selector, text: Regexp.compile(a.value), wait: wait_val)
+          end
+        when "check"
+          session.find(a.selector).set(true)
+        when "uncheck"
+          session.find(a.selector).set(false)
+        when "choose"
+          if a.options.nil?
+            session.find(a.selector).set(true)
+          else
+            session.find(a.selector + '[value="' + f[a.value].gsub('"', '\"') + '"]').set(true)
+          end
+        when "javascript"
+          session.driver.evaluate_script(a.value)
+        end
+      end
+
+      success = check_success session.text
+      form_fill_log(f, "done: #{success ? 'passing' : 'failing'} success criteria")
+
+      success_hash = {success: success}
+      success_hash[:screenshot] = self.class::save_screenshot_and_store_poltergeist(session) if !success
+      success_hash
+    rescue Exception => e
+      form_fill_log(f, "done: unsuccessful fill (#{e.class})")
+      Raven.extra_context(backtrace: e.backtrace)
+
+      message = {success: false, message: e.message, exception: e}
+      message[:screenshot] = self.class::save_screenshot_and_store_poltergeist(session)
+      message
+    ensure
+      session.driver.quit unless persist_session?
+    end
+  end
+
+  def check_success body_text
+    criteria = YAML.load(success_criteria)
+    criteria.each do |i, v|
+      case i
+      when "headers"
+        v.each do |hi, hv|
+          case hi
+          when "status"
+            # TODO: check status code
+          end
+        end
+      when "body"
+        v.each do |bi, bv|
+          case bi
+          when "contains"
+            unless body_text.include? bv
+              return false
+            end
+          end
+        end
+      end
+    end
+    true
+  end
+
+  def as_required_json o={}
+    as_json({
+      :only => [],
+      :include => {:required_actions => CongressMemberAction::REQUIRED_JSON}
+    }.merge o)
+  end
+
+
+  private
+
+  def form_fill_log(fields, message)
+    log_message = "#{bioguide_id} fill (#{[bioguide_id, fields].hash.to_s(16)}): #{message}"
+    Padrino.logger.info(log_message)
+
+    Raven.extra_context(fill_log: "") unless Raven.context.extra.key?(:fill_log)
+    Raven.context.extra[:fill_log] << message << "\n"
+  end
+
+  public
+
+  class_methods do
+    def crop_screenshot_from_coords screenshot_location, x, y, width, height
+      img = MiniMagick::Image.open(screenshot_location)
+      img.crop width.to_s + 'x' + height.to_s + "+" + x.to_s + "+" + y.to_s
+      img.write screenshot_location
+    end
+
+    def store_captcha_from_location location
+      c = CaptchaUploader.new
+      c.store!(File.open(location))
+      c.url
+    end
+
+    def store_screenshot_from_location location
+      s = ScreenshotUploader.new
+      s.store!(File.open(location))
+      s.url
+    end
+
+    def save_screenshot_and_store_poltergeist session
+      screenshot_location = random_screenshot_location
+      session.save_screenshot(screenshot_location, full: true)
+      url = store_screenshot_from_location screenshot_location
+      Raven.extra_context(screenshot: url)
+      File.unlink screenshot_location
+      url
+    end
+
+    def save_captcha_and_store_poltergeist session, x, y, width, height
+      screenshot_location = random_captcha_location
+      session.save_screenshot(screenshot_location, full: true)
+      crop_screenshot_from_coords screenshot_location, x, y, width, height
+      url = store_captcha_from_location screenshot_location
+      File.unlink screenshot_location
+      url
+    end
+
+    def random_captcha_location
+      "#{Padrino.root}/public/captchas/#{SecureRandom.hex(13)}.png"
+    end
+
+    def random_screenshot_location
+      "#{Padrino.root}/public/screenshots/#{SecureRandom.hex(13)}.png"
+    end
+  end
+end
